fix(cli): move DERIVED_DATA_DIR logic into DerivedDataLocator
Move the duplicated DERIVED_DATA_DIR environment variable resolution
from InspectBuildCommandService and InspectTestCommandService into
DerivedDataLocator.locate(for:) to centralize the logic.
